chore: refactor because files duplicated on merge
