+++ Sources/WebUI/Styles/Effects/Border/BorderStyleOperation.swift
