+++ Sources/WebUI/Styles/Effects/Border/BorderTypes.swift
