+++ Sources/WebUI/Styles/Effects/OutlineStyleOperation.swift
