+++ Sources/WebUI/Styles/Effects/RingStyleOperation.swift
